Init Coursera Course

+++ coursera-uc-san-diego/course1-algorithmic-toolbox/week1/APlusB.java
@@ -18,10 +18,10 @@ public class APlusB {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         
-        int a = scanner.nextInt();
-        int b = scanner.nextInt();  
+        long a = scanner.nextInt();
+        long b = scanner.nextInt();  
         
-        int sum = a + b;
+        long sum = a + b;
         
         System.out.println(sum);
